Fix to make expand task params as optional
Make expand task params as optional, so not all tasks need to provide it. 
Currently, if there is no expand field in tasks YAML, it fails to generate the DAG from YAML. Also using expand creates a task mapping in airflow which doesn't not work when you want to use a branch operator to skip certain tasks based on the condition.
Fix it here by making it optional.