Adds registration icon and links to Airflow references (#190)
"Airflow" has now been registered as a trademark in the US for the Apache Software Foundation. Whenever Airflow as a product is mentioned (especially in a commercial context), it must be referred "Apache Airflow®" and linked to https://airflow.apache.org. 

Given that the repo is owned by Astronomer, we will need to abide by the new rules. This commit updates all Airflow as a product references within the readme.