Fix weird test behaviour
The test was modifying a fixture, and this fixture is not in .gitignore.
The consequence is that if there were errors, the fixture was changed and developers would potentially commit an invalid fixture due to this.
This PR leverages the previous change introduced in #236 to use envvars for templating YML filepaths