Remove off looking start-date value in example_dag yaml config (#273)
The `example_dag` in the `example_dag_factory.yml` has a `start_date` 
value set to `2 days`, which seems unclear and problematic. It’s not
clear
what this value represents, and it appears to be causing issues with the
DAG, preventing it from running tasks. I confirmed this issue also
exists in
version `0.19.0`, so it’s not related to any changes made after that
version.
When I removed this questionable `start_date` value, the DAG ran its
tasks
successfully.

closes: #274