Update CODEOWNERS to track all files (#276)
The first argument in the CODEOWNERS file tells what all files to
monitor for changes. Since, we din't specify any file, the code review
requests are not getting sent. Specifying the wild card asterisk tells
it to monitor for all files in the repo & accordingly work for the
CODEOWNERS features such as sending review requests to the CODEOWNERS
upon opening pull requests.