Refactor dynamic task mapping implementation (#313)
Currently DAG Factory dynamic task mapping implementation relies in
internal Airflow implementation details (that it uses XComArg). If
Airflow changes the implementation, DAG factory will stop working.

With this PR, we're refactoring to use the `task.output` parameter that
is closer to the Airflow public API.

This was a suggestion by @ashb while we were discussing the overall DAG
Factory dynamic task implementation.