Remediated `default` behavior, added documentation (#378)
#295 outlined a documentation change that would better outline the
workflow for defining and using a "default" block when creating DAGs
with DAG Factory. In its most basic use, a `default` block at the top of
a YAML file would be used, followed by the DAG-definitions leveraging
these default values.

During the documentation and testing process, it was discovered that the
previously-expected behavior was not being implemented. This was due to
this code
(https://github.com/astronomer/dag-factory/blob/main/dagfactory/dagfactory.py#L136C1-L140C18),
which completely overrode `default_config`, rather than just the
`default_args` key-value pair.

To handle this, both a change to the logic above as well the appropriate
unit-tests were written. This change allowed for the behavior in #295 to
be completely implemented and documented.