Support dag-level arguments in global defaults (#480)
closes: https://github.com/astronomer/dag-factory/issues/462

This PR allows all DAG-level arguments to be set in `defaults.yml`.

Also, it merges `_load_yaml_config` back to `_load_dag_config`, so the
later one is the only function that handles the YAML file loading, with
custom constructor and Airflow 2 to 3 support. This is needed because
Airflow 2 to 3 logic is now required in the new `defaults.yml`.

---------

Co-authored-by: Young Li <gyli@uber.com>
Co-authored-by: Pankaj Singh <98807258+pankajastro@users.noreply.github.com>