Resolve 'credential persistence through GitHub Actions artifacts' warnings from Zizmor (#494)
The persist-credentials option in the actions/checkout action determines
whether the Git credentials used to fetch the repository are left in the
Git configuration after the checkout step. Setting `persist-credentials`
to `true` (the default) means the credentials will remain, allowing
subsequent steps in the workflow to perform authenticated Git operations
without re-authentication. Setting it to false removes the credentials,
which is useful for security when you don't want later steps to have
access to the credentials.

scan results from this PR:
https://github.com/astronomer/dag-factory/security/code-scanning?query=is%3Aopen+pr%3A494

vs

on main:
https://github.com/astronomer/dag-factory/security/code-scanning  (this will change after merging this PR, but can be compared while the PR is still open)


related:
https://github.com/astronomer/oss-integrations-private/issues/156