Fix the Airflow version condition check to parse inlets/outlets syntax according to the dataset (#485)
We should be able to parse inlets/outlets syntax according to the
dataset for Airflow < 3.0.0, because in Airflow 3, the dataset has been
replaced with an asset. This PR fixes the wrong condition check for
Airflow version for parsing in inlets/outlets for datasets

This PR also add an example DAG for Asset.

Before this PR, DAG below will not work.

```yaml
default:
  default_args:
    start_date: 2025-01-01
  catchup: false

producer_dag:
  schedule: "@daily"
  tasks:
    produce_data:
      operator: "airflow.providers.standard.operators.python.PythonOperator"
      python_callable: utils.generate_data
      outlets:
        - __type__: airflow.sdk.Asset
          uri: "file:///$AIRFLOW_HONE/data.csv"
          name: "data_asset"

consumer_dag:
  schedule:
    - __type__: airflow.sdk.Asset
      uri: "file:///$AIRFLOW_HONE/data.csv"
      name: "data_asset"
  tasks:
    consume_data:
      operator: "airflow.providers.standard.operators.bash.BashOperator"
      bash_command: "echo 'Asset was updated, running DAG!'"
```