Update pyproject.toml to Sync Test Versions with CI/CD Pipeline (#553)
### Description

**Update Python and Airflow versions in `pyproject.toml`**: The file has
been updated to mirror the configurations in the CI pipeline. This
ensures that versions like **Python 3.11** and **Airflow 2.4** are
excluded in both local and CI environments.

### Changes

* Excluded Python 3.11 with Airflow versions 2.4, 2.5, and 2.6: These
combinations are excluded due to compatibility issues with Python 3.11.
* Excluded Python 3.12 with Airflow versions 2.4, 2.5, 2.6, 2.7, and
2.8: Official support for Python 3.12 and corresponding constraints.txt
files are available only for Apache Airflow versions >= 2.9.0.
Therefore, Python 3.12 is excluded for Airflow versions prior to 2.9.0.

### Related Issues:

#552