Fix incorrect `load_yaml_dags` parameter usage (#555)
### Description
The `load_yaml_dags` function in `dev/dags/example_object_storage.py`
was incorrectly provided with a single file path for the `dags_folder`
parameter. This parameter is intended for folder paths and was causing
the function to skip the recursive scanning of YAML files. As a result,
the associated DAG was not being loaded and therefore not included in
the integration tests.

### Changes
change from `dags_folder` to `config_filepath` parameter to ensure
`example_object_storage` could be included in integration testing.

### Related Issue
Closes: #554