Deprecating sla_miss_callback for Airflow >= 3.1.0 (#586)
I was notified of the following issue. This PR aims to fix that:

> This might be expected, but I wanted to check if you're aware of it. I
tried dag-factory 1.0.0 with the latest Airflow 3.1 nightly
:airflow_spin:.
> 
> Even a simple DAG definition fails with:
> 
> ```
> TypeError: DAG.__init__() got an unexpected keyword argument
'sla_miss_callback'
> The issue is that dagbuilder.py sets the parameter:
> dag_kwargs["sla_miss_callback"] = dag_params.get("sla_miss_callback",
None)
> ```
> 
> every time a DAG is instantiated. The parameter was [deprecated in
3.0](https://airflow.apache.org/docs/task-sdk/stable/api.html#airflow.sdk.DAG)
and removed in 3.1. I assume this will be addressed once 3.1 is
released, but I couldn't find an open issue yet, so I wanted to flag it.